face-api-proto/face-api-server/requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Assets==0.12
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
python-dotenv==0.10.3
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
webassets==0.12.1
Werkzeug==0.16.0
wrapt==1.11.2
